updated_datasets

+++ data_creation_code/create_title_excels.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import os
 
-def split(file = None, data_dir = r'../data/publications/orcid', export_dir = r'../data/publications/orcid'):
+def split(file = None, data_dir = r'../data/publications/orcid', export_dir = r'../data/publications/orcid', overwriteIDs = False):
     #get files' names
     if file == None:
         directories = os.listdir(data_dir)
@@ -18,20 +18,21 @@ def split(file = None, data_dir = r'../data/publications/orcid', export_dir = r'
         name = 'banana'
         
         # create ID's
-        cnt = 1
-        for fullname in df.name.values:
-            fullname = fullname.replace(' ', '_')
-            if fullname == name:
-                cnt += 1
-            else:
-                cnt = 1
-                name = fullname
-            id = fullname + '_' + str(cnt)
-            id_list.append(id)
-        df['id'] = id_list
-        
-        # overwrite with ids
-        df.to_csv(rf'../data/publications/orcid/{file}')
+        if overwriteIDs:
+            cnt = 1
+            for fullname in df.name.values:
+                fullname = fullname.replace(' ', '_')
+                if fullname == name:
+                    cnt += 1
+                else:
+                    cnt = 1
+                    name = fullname
+                id = fullname + '_' + str(cnt)
+                id_list.append(id)
+            df['id'] = id_list
+            
+            # overwrite with ids
+            df.to_csv(rf'../data/publications/orcid/{file}')
         
         
         # save
